change priority type to byte

+++ api/src/main/kotlin/com/dustrean/api/event/CoreListener.kt
@@ -1,5 +1,5 @@
 package com.dustrean.api.event
 
 annotation class CoreListener(
-    val priority: Int = 0
+    val priority: Byte = 0
 )
